Extract legacy database into a separate file
